move tree/model.py to tree/treemodel.py
