+++ rootpy/tree/treemodel.py
